ps4
Practicing thinking about problems in a recursive fashion, taking advantage of the idea that one can reduce the problem to a simpler version of the same problem. Also, thinking about problems in terms of classes, each instance of which contains specific attributes as well as methods for manipulating them using object-oriented programming to write a Caesar/shift cipher.